Move Alpine definition to the python3 key.
The Alpine definition for this key is for python3 and thus belongs under
the python3 variant of this key.

Because on Debian and Ubuntu the same package provides support for both
python2 and python3 it's possible that this python2-by-convention key is
being used where a python3 dependency is expected. This is something we
should update in sources or unify the two keys and mandate that
platforms which package python2 and python3 support separately resolve
both with one key.