Update the FreeBSD dependencies for Rosdep. (#22335)
* Update the FreeBSD dependencies for Rosdep.

Since the pkg backend is  package system that is now used on FreeBSD
some port names have been changed. Since the rosdep and friends have
been added to the FreeBSD ports tree, it is a good thing to keep this
up to date. There may still be holes, but it is complete enough to
build the base comms and desktop releases of melodic.

Also document where FreeBSD packages can be found.