Mechanism to check rosdep rules against package repositories (#28902)
This change adds a small system for checking rosdep rules against the
current package repositories for supported platforms.

It currently supports three platforms:
- Ubuntu
- Fedora
- RHEL

During a typical rosdep test invocation, it uses a similar mechanism as
the one which verifies URLs to detect which lines were added or modified
by the latest change, and reviews only those rules.

The algorithm is optimized such that the sources are listed in the order
they are checked for the package, and parsing of the repository metadata
begins before the download is even completed. This means that presence
in the platform repositories can be determined very quickly.

If multiple package names were added or modified, the algorithm
first checks the data that has already been parsed and then continues
parsing metadata until the package is found or the repository list is
exhausted.

To review the entire rosdep database and not only the rules that were
changed in the most recent commit, it is also possible to invoke the
test as a python module. For example:

`PYTHONPATH=test python3 -m rosdep_repo_check`

This system is currently identifying many missing packages throughout
rosdep. We should review these issues at some point, but this system
should help to prevent more problematic rules from being added.

Co-authored-by: Steven! Ragnarök <nuclearsandwich@users.noreply.github.com>